chore: organize http handlers
They're currently randomly in a bunch of different files. This cleans up
the handler functions to be in the file of the type they return.
